2018.3.5

+++ app/views/api_v1/homeuser/__init__.py
@@ -1,7 +1,7 @@
 import datetime
 
 from flask import g, request
-from flask_restplus import Namespace, Resource
+from flask_restplus import Namespace, Resource, abort
 from sqlalchemy import DateTime, and_
 
 from app.ext import db
@@ -27,12 +27,13 @@ class HomeUsersView(Resource):
     @page_range()
     def get(self):
         list = HomeUser.query
-        home=Home.query.filter(Home.admin_user_id==g.user.id)
-        if g.role.name in ['admin','superadmin']:
-            return list, 200
-        else:
-            return list.filter(
-              HomeUser.home_id.in_(i.id for i in home)).filter(HomeUser.if_confirm == True), 200
+        home = Home.query.filter(Home.admin_user_id == g.user.id).filter(Home.disabled == False).all()
+        if home:
+            if g.role.name in ['admin', 'superadmin']:
+                return list, 200
+            else:
+                return list.filter(HomeUser.home_id.in_(i.id for i in home)).filter(HomeUser.if_confirm == True),
+        else: abort(404, message='家庭不存在')
 
 
 @api.route('/<homeid>/')
@@ -42,9 +43,10 @@ class HomeUserView1(Resource):
     @role_require(['homeuser'])
     @api.response(200, 'ok')
     def post(self, homeid):
-        try:
+        home = Home.query.filter(Home.disabled == False).filter(Home.id == homeid).first()
+        if home:
             homeuser = HomeUser.query.filter(HomeUser.home_id == homeid)
-            if Home.query.get_or_404(homeid):
+            if Home.query.filter(Home.disabled == False).filter(Home.id == homeid).first():
                 if g.user.id not in [i.user_id for i in homeuser]:
                     homeuser = HomeUser()
                     homeuser.home_id = homeid
@@ -54,25 +56,23 @@ class HomeUserView1(Resource):
                     return '申请成功', 200
                 else:
                     return '您已经是该家庭成员', 201
+        else:return '家庭不存在',401
 
-        except:return '该家庭未创建', 401
-
-
-
-    @page_format(code=0,msg='ok')
+    @page_format(code=0, msg='ok')
     @api.doc('显示家庭申请')
     @api.header('jwt', 'JSON Web Token')
     @role_require(['homeuser'])
-    @api.marshal_with(homeuser_model,as_list=True)
+    @api.marshal_with(homeuser_model, as_list=True)
     @api.response(200, 'ok')
     @page_range()
-    def get(self,homeid):
-       home=Home.query.get_or_404(homeid)
-       if g.user.id==home.admin_user_id:
-           homeuser = HomeUser.query.filter(HomeUser.if_confirm==False).filter( HomeUser.home_id==homeid )
-           return homeuser,200
-       else: pass
-
+    def get(self, homeid):
+       home = Home.query.filter(Home.disabled == False).filter(Home.id == homeid).first()
+       if home:
+           if g.user.id == home.admin_user_id:
+               homeuser = HomeUser.query.filter(HomeUser.if_confirm == False).filter(HomeUser.home_id == homeid)
+               return homeuser, 200
+           else: return '权限不足', 200
+       else:abort(404, message='家庭不存在')
 
 
 @api.route('/<homeid>/<userid>/')
@@ -82,33 +82,36 @@ class HomeUserView2(Resource):
     @role_require(['homeuser'])
     @api.response(200, 'ok')
     @user_require
-    def put( self,homeid,userid):
-        home = Home.query.get_or_404(homeid)
-        homeuser = HomeUser.query.filter(and_(HomeUser.home_id == homeid ,HomeUser.user_id == userid)).first()
-        homeuser.if_confirm = True
-        homeuser.confirm_time = datetime.datetime.now()
-        if g.user.id == home.admin_user_id:
-            db.session.commit()
-            return '绑定成功', 200
-        else:
-            return '权限不足', 201
-
+    def put(self, homeid, userid):
+        home = Home.query.filter(Home.disabled == False).filter(Home.id == homeid).first()
+        if home:
+            homeuser = HomeUser.query.filter(and_(HomeUser.home_id == homeid, HomeUser.user_id == userid)).first()
+            homeuser.if_confirm = True
+            homeuser.confirm_time = datetime.datetime.now()
+            if g.user.id == home.admin_user_id:
+                db.session.commit()
+                return '绑定成功', 200
+            else:
+                return '权限不足', 201
+        else: return '家庭不存在', 201
 
     @api.doc('删除家庭成员记录')
     @api.header('jwt', 'JSON Web Token')
-    @role_require(['homeuser','admin','superadmin'])
+    @role_require(['homeuser', 'admin', 'superadmin'])
     @api.response(200, 'ok')
     def delete(self, homeid, userid):
-        home = Home.query.get_or_404(homeid)
-        homeuser = HomeUser.query.filter(HomeUser.home_id == homeid).filter(HomeUser.user_id == userid).first()
-        db.session.delete(homeuser)
-        if g.role.name=='homeuser':
-            if g.user.id == home.admin_user_id:
-                db.session.commit()
-                return '删除成功', 200
+        home = Home.query.filter(Home.disabled == False).filter(Home.id == homeid).first()
+        if home:
+            homeuser = HomeUser.query.filter(HomeUser.home_id == homeid).filter(HomeUser.user_id == userid).first()
+            db.session.delete(homeuser)
+            if g.role.name == 'homeuser':
+                if g.user.id == home.admin_user_id:
+                    db.session.commit()
+                    return '删除成功', 200
+                else:
+                    return '权限不足', 201
             else:
-                return '权限不足', 201
-        else:
-            db.session.commit()
-            return'删除成功',200
+                db.session.commit()
+                return'删除成功', 200
+        else:return '家庭不存在', 201
